Some cleanup, including moving jit fns to separate file.
